pyproject.toml
[project]
name = "glofas-anadyr"
version = "0.1.0"
description = "GloFAS data download script"
requires-python = "==3.12.*"
dependencies = [
  "boto3>=1.42.9",
  "cdsapi == 0.7.7",
  "fastparquet>=2024.11.0",
  "hydroeval>=0.1.0",
  "numpy== 2.3.5",
  "pandas>=2.3.3",
  "pyarrow>=22.0.0",
  "python-dotenv>=1.0.0",
  "xarray>=2025.12.0",
  "xsdba == 0.5.0",
]
